Separated the search bar into a component itself
