+++ my-app/src/components/SearchBar/index.jsx
@@ -1,9 +1,41 @@
 import React from 'react';
+import { useState } from 'react';
+import axios from 'axios';
 import './style.css';
 
-const SearchBar = () => {
+const SearchBar = (props) => {
+
+    const [textInput,setTextInput] = useState("");
+
+    const handleChange = (e) => {
+        setTextInput(e.target.value);
+    }
+
+    const handleSearch = () => {
+        const query = textInput;
+        let accessTokenBearer = `Bearer ${props.accessToken}`;
+        const BASE_URL = "https://api.spotify.com/v1/search?q="
+        console.log(accessTokenBearer)
+        const getSpotifySearch = async() => {
+            try {
+                const response = await axios.get(`${BASE_URL}${query}&type=album&limit=30`,{
+                    headers: {
+                        'Authorization': accessTokenBearer
+                    }
+                })
+                props.getData(response.data);
+            } catch(error){
+                console.error(error);
+            }
+        }
+        getSpotifySearch();
+    }
+
     return(
-        <input className="searchBar" type="text"></input>
+        <div>
+            <input type="text" value={textInput} onChange={handleChange}></input>
+            <button onClick={handleSearch}>Search!</button>
+        </div>
     );
 }
